Week2/airflow/requirements.txt
apache-airflow-providers-google
pyarrow
google-cloud-storage
